requirements.txt
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
db-dtypes==1.0.3
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
proto-plus==1.22.0
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
toml==0.10.2
urllib3==1.26.12
